+++ src/help.spec.ts
@@ -1,53 +1,8 @@
 import { BBHelpClient } from './help';
+import { MockHelpWidget } from './mocks/mock-help-widget';
 
 describe('BBHelpClient', () => {
   let mockHelpWidget: any;
-
-  class MockHelpWidget {
-    public close() {
-      return;
-    }
-
-    public load(helpConfig?: any) {
-      return;
-    }
-
-    public open(helpKey?: string) {
-      return;
-    }
-
-    public toggleOpen() {
-      return;
-    }
-
-    public disableWidget() {
-      return;
-    }
-
-    public enableWidget() {
-      return;
-    }
-
-    public ready() {
-      return Promise.resolve();
-    }
-
-    public setHelpKeyToDefault() {
-      return;
-    }
-
-    public setCurrentHelpKey() {
-      return;
-    }
-
-    public getCurrentHelpKey() {
-      return;
-    }
-
-    public getWhatsNewRevision() {
-      return;
-    }
-  }
 
   beforeEach(() => {
     mockHelpWidget = new MockHelpWidget();
