+++ src/mocks/mock-help-widget.ts
@@ -1,0 +1,41 @@
+export class MockHelpWidget {
+  public close() {
+    return;
+  }
+
+  public load(helpConfig?: any) {
+    return;
+  }
+
+  public open(helpKey?: string) {
+    return;
+  }
+
+  public toggleOpen() {
+    return;
+  }
+
+  public disableWidget() {
+    return;
+  }
+
+  public enableWidget() {
+    return;
+  }
+
+  public ready() {
+    return Promise.resolve();
+  }
+
+  public setHelpKeyToDefault() {
+    return;
+  }
+
+  public setCurrentHelpKey() {
+    return;
+  }
+
+  public getWhatsNewRevision() {
+    return;
+  }
+}
